+++ tests/test_scripts/ms/test_ms_germline_risk.py
@@ -141,7 +141,7 @@ def test_variant_edge_cases(lightweight_test_run, tmp_path, deduped_bam, deduped
 
     # Copy snakemake files to temporary directory
     shutil.copy("Snakefile", tmp_path / "Snakefile")
-    shutil.copytree("scripts", tmp_path / "scripts")
+    shutil.copytree("rule_scripts", tmp_path / "rule_scripts")
     shutil.copytree("rules", tmp_path / "rules")
     shutil.copytree("tmp/downloads", tmp_path / "tmp/downloads")
     shutil.copytree("tests/data/lightweight_test_run/config", tmp_path / "tests/data/lightweight_test_run/config")